- Added field validation for git credentials control. Error messages are displayed underneath the field that has an error. - Git export wizard next button is disabled when field validation fails. - Layout changes to repository information section of the git credentials control.
Issue: The git credentials control repository properties border title css markup needs work so that the border does not appear behind the title.
Issue: Since ng-message needs an ng-model to work, and input form controls that have a type of file do not have an ng-model, the validation logic does not work when those input controls are involved.
Issue: Seems like the export wizard height should be a bit larger. I think there is a max-height somewhere that controls this.